refactor: move home page script to pages directory
